[refactor] : RacingGameUsecase 인터페이스 분리 및, app 패키지 생성
