+++ src/main/java/racingcar/app/input/ConsoleInput.java
@@ -1,4 +1,4 @@
-package racingcar.input;
+package racingcar.app.input;
 
 import java.util.Arrays;
 import java.util.List;
