+++ src/main/java/racingcar/controller/RacingGamePlayController.java
@@ -6,21 +6,22 @@ import racingcar.controller.request.RacingGamePlayControllerRequest;
 import racingcar.controller.response.RacingGamePlayControllerResponse;
 import racingcar.randommove.RandomMover;
 import racingcar.usecase.RacingGamePlayUsecase;
+import racingcar.usecase.RacingGamePlayable;
 import racingcar.usecase.request.RacingGamePlayRequest;
 import racingcar.usecase.request.RacingGamePlayRequest.CarRequest;
 import racingcar.usecase.response.RacingGamePlayResponse;
 
 public class RacingGamePlayController {
 
-    private final RacingGamePlayUsecase racingGamePlayUsecase;
+    private final RacingGamePlayable racingGamePlayable;
     private static final RandomMover randomMover = RandomMover.newInstance();
 
     public RacingGamePlayController(RacingGamePlayUsecase racingGamePlayUsecase) {
-        this.racingGamePlayUsecase = racingGamePlayUsecase;
+        this.racingGamePlayable = racingGamePlayUsecase;
     }
 
     public RacingGamePlayControllerResponse playRacingGame(RacingGamePlayControllerRequest request) {
-        RacingGamePlayResponse playResult = racingGamePlayUsecase
+        RacingGamePlayResponse playResult = racingGamePlayable
             .play(request.getRound(), getRacingGamePlayRequest(request.getCarNames()));
 
         return new RacingGamePlayControllerResponse(playResult);
